refactor: enhance database schema and queries for werewolf game and lobby, updating block height types to BIGINT and consolidating type exports

+++ packages/client/database/src/migrations/werewolf_game.sql
@@ -1,19 +1,19 @@
 -- Tracks the state of each game round for vote timeout enforcement
 CREATE TABLE werewolf_round_state (
-  game_id INTEGER NOT NULL,
+  game_id BIGINT NOT NULL,
   round INTEGER NOT NULL,
   phase TEXT NOT NULL,
   alive_count INTEGER NOT NULL DEFAULT 0,
   votes_submitted INTEGER NOT NULL DEFAULT 0,
-  round_started_block INTEGER NOT NULL,
-  timeout_block INTEGER,
+  round_started_block BIGINT NOT NULL,
+  timeout_block BIGINT,
   resolved BOOLEAN NOT NULL DEFAULT FALSE,
   PRIMARY KEY (game_id, round, phase)
 );
 
 -- Snapshot of alive player indices at the start of each round (used to determine who to punish on timeout)
 CREATE TABLE werewolf_alive_snapshot (
-  game_id INTEGER NOT NULL,
+  game_id BIGINT NOT NULL,
   round INTEGER NOT NULL,
   phase TEXT NOT NULL,
   player_idx INTEGER NOT NULL,
@@ -23,9 +23,9 @@ CREATE TABLE werewolf_alive_snapshot (
 -- Pending player punishments to be dispatched to the Midnight contract by the admin
 CREATE TABLE werewolf_pending_punishments (
   id SERIAL PRIMARY KEY,
-  game_id INTEGER NOT NULL,
+  game_id BIGINT NOT NULL,
   player_idx INTEGER NOT NULL,
   reason TEXT NOT NULL DEFAULT 'vote_timeout',
-  created_at_block INTEGER NOT NULL,
+  created_at_block BIGINT NOT NULL,
   executed BOOLEAN NOT NULL DEFAULT FALSE
 );
